feat: add usePublicationAttachmentStore
